+ Renamed domain to pack + Seperated model stuff from pack into new model class + Fixed vtp writing for 2d + VTP writing works correctly

+++ writer.h
@@ -19,7 +19,7 @@ class Writer {
       std::vector<IntField*> intFields,
       std::vector<DoubleField*> doubleFields,
       int dim,
-      int numParticles
+      long numParticles
     )=0;
 };
